langchain_examples/requirements.txt
langchain-openai
openai
python-dotenv
